requirements.txt
Django
django-log-request-id
python-dotenv
setproctitle 
gunicorn
